refactoring to organize existing code
